requirements.txt
asgiref==3.11.1
Django==6.0.2
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0